// src/app/models/Response/Venta/TB_Cliente/Vw-Cliente.ts
export class VWCliente{
    idCliente: number = 0;
    nombrePersona: string = "";
    tipoPersona: string = "";
    tipoDocumento: string = "";
    numeroDocumento: string = "";
    telefono: string = "";
    codigoUbigeo: string | null = "";
    direccion: string | null = "";
    fechaNacimiento: string | null = "";
    estado: boolean = true;
}